+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   - resolve -> use graphql resolvers
   - @value -> use @field.instruct directive
 - Removed toId Transform -> use @field(id: '')
-- Removed Model.tform() -> use Model.hydrate(data, context)
+- Removed Model.tform() -> use Model.shapeObject(shape, data)
 - Removed Resolver.toResultSet() -> ? TBD ?
 - Removed Transformer + Rule -> use Pipeline
   - Removed many pre-defined rules + transformers